fix: make the actors bundle build on crates.io
This commit:

1. Moves the "bundle" to the workspace root. That way the generated
package includes the wasm profile.
2. Un-excludes the Cargo.lock file and adds a dummy example to force
its inclusion in the crates.io package. This ensures that we use the
correct dependency versions, and avoids "dirtying" the build environment
when we build the bundle.
3. Builds the wasm bundle with "--locked".

The _alternative_ was to do what substrate's wasm builder does:
dynamically generate a "fake" wasm build package/workspace and build
that. But this way was simpler and had the added bonus of using the same
lockfile everywhere.